requirements.txt
streamlit>=1.30.0
langgraph>=0.2.0
langchain-anthropic>=0.2.0
anthropic>=0.40.0
openai>=1.50.0
supabase>=2.0.0
pgvector>=0.3.0
ffmpeg-python>=0.2.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
typing-extensions>=4.9.0
pillow>=10.0.0
